feat(dataplane): expose structs and methods useful to parse exported samples
