+++ protos/dataplane.proto
@@ -1,5 +1,5 @@
 syntax = "proto3";
-option go_package = "github.com/neblic/platform/sampler/protos";
+option go_package = "github.com/neblic/platform/dataplane/protos";
 
 message IntNumType { int64 count = 1; }
 message UIntNumType { int64 count = 1; }
